+++ KioChess/Engine/Strategies/AB/NegaMaxMemoryStrategy.cs
@@ -6,13 +6,13 @@ namespace Engine.Strategies.AB;
 
 public class NegaMaxMemoryStrategy : MemoryStrategyBase
 {
-    public NegaMaxMemoryStrategy(short depth, IPosition position, TranspositionTable table = null) : base(depth, position, table)
+    public NegaMaxMemoryStrategy(int depth, IPosition position, TranspositionTable table = null) : base(depth, position, table)
     {
         InitializeSorters(depth, position, MoveSorterProvider.GetSimple(position));
     }
 
     protected override StrategyBase CreateSubSearchStrategy()
     {
-        return new NegaMaxMemoryStrategy((short)(Depth - SubSearchDepth), Position);
+        return new NegaMaxMemoryStrategy(Depth - SubSearchDepth, Position);
     }
 }
