+++ KioChess/Engine/Strategies/AB/NegaMaxStrategy.cs
@@ -5,13 +5,13 @@ namespace Engine.Strategies.AB;
 
 public class NegaMaxStrategy : StrategyBase
 {
-    public NegaMaxStrategy(short depth, IPosition position) : base(depth, position)
+    public NegaMaxStrategy(int depth, IPosition position) : base(depth, position)
     {
         InitializeSorters(depth, position, MoveSorterProvider.GetSimple(position));
     }
 
     protected override StrategyBase CreateSubSearchStrategy()
     {
-        return new NegaMaxStrategy((short)(Depth - SubSearchDepth), Position);
+        return new NegaMaxStrategy(Depth - SubSearchDepth, Position);
     }
 }
