+++ KioChess/Engine/Strategies/Lmr/LmrDeepStrategy.cs
@@ -6,7 +6,7 @@ namespace Engine.Strategies.Lmr;
 
 public class LmrDeepStrategy : LmrStrategyBase
 {
-    public LmrDeepStrategy(short depth, IPosition position, TranspositionTable table = null) : base(depth, position, table)
+    public LmrDeepStrategy(int depth, IPosition position, TranspositionTable table = null) : base(depth, position, table)
     {
     }
